Stabilizing v0.4 (#39)
Stabilizing v0.4

New Features
- touching beacon and device to identify and (de-)activate beacon
- OTA functionality completely implemented and documented
- Function to remove bonded devices and to re-negotiate with device

Bugfixes
- error handling if GATT connection not established
- exception handling due to not completed downloads or cancel pressed
- mqtt counter corrected
- ingroring bottom/top range values from temperature/humidity sensor
- WDT rewritten to initiate reboot on correct reason (too long not seen, etc.)

Optimization
- display cleanup
- web server status table layout and data field extension

Technical/Monitoring
- send free heap, uptime, beacon last seen, MQTT last send information for monitoring purposes
- OTA trigger in web server available
- reboot trigger in web server available

Documentation
- Worked on documentation

+++ main/ble_mqtt.h
@@ -14,8 +14,8 @@ extern EventGroupHandle_t mqtt_evg;
 #define  MQTT_CONNECTED_BIT      (BIT0)
 
 
-extern uint16_t mqtt_packets_send;
-extern uint16_t mqtt_packets_fail;
+extern uint32_t mqtt_packets_send;
+extern uint32_t mqtt_packets_fail;
 
 bool send_to_mqtt(uint8_t idx, uint16_t maj, uint16_t min, float temp, float humidity, uint16_t battery, int8_t rssi);
